requirements.txt
flask==3.1.0
flask-cors==5.0.0
PyJWT==2.8.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
